FINISHED - EVERYTHING WORKS AND TESTED. ADDED ALL LOOKUP VALUES, CLEANED UP THE CONSOLE LOGS AND COMMENTS.
Only map that doesnt have lookup values is vertigo. It is also the only other map apart from nuke that has a split and when i tested with mirage values, the split functioned correctly so all you would have to do is make a lookup table and it should work just like nuke.